% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashgraph.R
\name{fg.get.sparse.matrix}
\alias{fg.get.sparse.matrix}
\title{Load a sparse matrix to FlashGraphR.}
\usage{
fg.get.sparse.matrix(fg)
}
\description{
Load a sparse matrix to FlashGraphR from FlashGraph.
}
\details{
`fg.get.sparse.matrix' gets a FlashMatrix sparse matrix that references
a matrix represented by a FlashGraph object.
}
\examples{
fg <- fg.load.graph("graph.adj", "graph.index")
fm <- fg.get.sparse.matrix(fg)
}
