% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashgraph.R
\name{fg.lcc}
\alias{fg.lcc}
\alias{fg.get.lcc}
\title{Get the largest connected component in a graph}
\usage{
fg.get.lcc(graph, mode = c("weak", "strong"))
}
\arguments{
\item{graph}{The FlashGraph object}

\item{mode}{Character string, either "weak" or "strong". For directed
graphs "weak" implies weakly, "strong" strongly c
components to search. It is ignored for undirected graphs.}
}
\value{
a FlashGraph object that contains the largest connected component in
        the graph.
}
\description{
Get the largest (weakly or strongly) connected component in a graph.
}
\details{
For an undirected graph, this function returns the largest connected component
and ignores the argument `mode'.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}
