% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashgraph.R
\name{fg.print.graph}
\alias{fg.print.graph}
\title{Print a graph into a file as an edge list.}
\usage{
fg.print.graph(fg, file, delim = "\\t", type = "")
}
\arguments{
\item{fg}{the FlashGraph object}

\item{file}{a string of the output file name.}
}
\description{
Print a graph in the FlashGraph format into a file as an edge list.
}
