% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashgraph.R
\name{fg.subgraph}
\alias{fg.subgraph}
\alias{fg.fetch.subgraph}
\title{Generate an induced subgraph}
\usage{
fg.fetch.subgraph(graph, vertices, name = paste(graph$name, "-sub", sep = ""),
  compress = TRUE)
}
\arguments{
\item{graph}{The FlashGraph object}

\item{vertices}{A numeric vector that contains the ids of vertices in
the induced subgraph.}

\item{name}{The name of the FlashGraph object.}

\item{compress}{This indicates whether to remove empty vertices in
the generated subgraph.}
}
\value{
`fg.fetch.subgraph' returns a FlashGraph object
}
\description{
Generate an induced subgraph that contains the specified vertices from
the input given graph.
}
\details{
`fg.fetch.subgraph' generates the induced subgraph and returns
a FlashGraph object.
}
\author{
Da Zheng <dzheng5@jhu.edu>
}
